// src/shared/types/Car.ts
export interface Car {
    id: string;
    carModelId: string;
    year: number;
    description: string;
    color: string;
    mileage: number;
    ownersCount: number;
    transmissionType: string;
    bodyType: string;
    enginePower: string;
    engineType: string;
    driveType: string;
    engineCapacity: string;
    steeringPosition: string;
    seatsCount: number;
    price: number;
    isAvailable: boolean;
    createdAt: string;
  }